bug #17195 - allow both text and integer orderseq values to be returned to the pool, with text that cannot be coerced to ints being ignored for now. this is a revision of the fix for issue #17119.

+++ xtupleserver/trunk/dbscripts/triggers/apopen.sql
@@ -30,7 +30,7 @@ BEGIN
     END IF;
 
      --- clear the number from the issue cache
-    PERFORM clearNumberIssue('APMemoNumber', NEW.apopen_docnumber::INTEGER);
+    PERFORM clearNumberIssue('APMemoNumber', NEW.apopen_docnumber);
   END IF;
   
   IF (TG_OP = 'UPDATE') THEN
